Type the HSD object allocator pool as addresses
obj_heap held its pool bounds as u32, and HSD_ObjAllocAddFree threaded the free list through (void**) byte offsets. The bounds are now uintptr_t, the pool arithmetic stays in that domain, and the free list is threaded through a typed HSD_ObjAllocLink, which also accounts for the stack space the u8 _[4] pad was standing in for.

+++ src/sysdolphin/baselib/objalloc.h
@@ -6,8 +6,8 @@
 #include <sysdolphin/baselib/debug.h>
 
 typedef struct _objheap {
-    u32 top;
-    u32 curr;
+    uintptr_t top;
+    uintptr_t curr;
     u32 size;
     u32 remain;
 } objheap;
